Time: 0 ms (100.00%) | Memory: 30.5 MB (74.87%) - LeetSync

+++ 3263-divide-an-array-into-subarrays-with-minimum-cost-i/divide-an-array-into-subarrays-with-minimum-cost-i.cpp
@@ -1,9 +1,9 @@
 class Solution {
 public:
     int minimumCost(vector<int>& nums) {
-        long long sum = nums[0];
+        int sum = nums[0];
         int min_idx = -1;
-        long long min_value = INT_MAX;
+        int min_value = INT_MAX;
         for(int i = 1; i<nums.size(); i++){
             if(nums[i] < min_value){
                 min_value = nums[i];
@@ -20,6 +20,6 @@ public:
             }
         }
         sum+=min_value;
-        return (int)sum;
+        return sum;
     }
 };
